refactor 4 mutations to sdl
