--- packages/server/graphql/mutations/renameReflectTemplatePrompt.ts
@@ -1,83 +1,0 @@
-import {GraphQLID, GraphQLNonNull, GraphQLString} from 'graphql'
-import {SubscriptionChannel} from 'parabol-client/types/constEnums'
-import getRethink from '../../database/rethinkDriver'
-import getKysely from '../../postgres/getKysely'
-import {getUserId, isTeamMember} from '../../utils/authorization'
-import publish from '../../utils/publish'
-import standardError from '../../utils/standardError'
-import {GQLContext} from '../graphql'
-import RenameReflectTemplatePromptPayload from '../types/RenameReflectTemplatePromptPayload'
-
-const renameReflectTemplatePrompt = {
-  description: 'Rename a reflect template prompt',
-  type: RenameReflectTemplatePromptPayload,
-  args: {
-    promptId: {
-      type: new GraphQLNonNull(GraphQLID)
-    },
-    question: {
-      type: new GraphQLNonNull(GraphQLString)
-    }
-  },
-  async resolve(
-    _source: unknown,
-    {promptId, question}: {promptId: string; question: string},
-    {authToken, dataLoader, socketId: mutatorId}: GQLContext
-  ) {
-    const r = await getRethink()
-    const pg = getKysely()
-    const now = new Date()
-    const operationId = dataLoader.share()
-    const subOptions = {operationId, mutatorId}
-    const prompt = await dataLoader.get('reflectPrompts').load(promptId)
-    const viewerId = getUserId(authToken)
-
-    // AUTH
-    if (!prompt || prompt.removedAt) {
-      return standardError(new Error('Prompt not found'), {userId: viewerId})
-    }
-    if (!isTeamMember(authToken, prompt.teamId)) {
-      return standardError(new Error('Team not found'), {userId: viewerId})
-    }
-
-    // VALIDATION
-    const {teamId, templateId} = prompt
-    const trimmedQuestion = question.trim().slice(0, 100)
-    const normalizedQuestion = trimmedQuestion || 'Unnamed Prompt'
-
-    const prompts = await dataLoader.get('reflectPromptsByTemplateId').load(templateId)
-    const allPrompts = prompts.filter(({removedAt}) => !removedAt)
-    if (allPrompts.find((prompt) => prompt.question === normalizedQuestion)) {
-      return standardError(new Error('Duplicate question template'), {userId: viewerId})
-    }
-
-    // RESOLUTION
-    await Promise.all([
-      r
-        .table('ReflectPrompt')
-        .get(promptId)
-        .update({
-          question: normalizedQuestion,
-          updatedAt: now
-        })
-        .run(),
-      pg
-        .updateTable('ReflectPrompt')
-        .set({question: normalizedQuestion})
-        .where('id', '=', promptId)
-        .execute()
-    ])
-    dataLoader.clearAll('reflectPrompts')
-    const data = {promptId}
-    publish(
-      SubscriptionChannel.TEAM,
-      teamId,
-      'RenameReflectTemplatePromptPayload',
-      data,
-      subOptions
-    )
-    return data
-  }
-}
-
-export default renameReflectTemplatePrompt
