+++ packages/server/graphql/public/types/ReflectTemplatePromptUpdateDescriptionPayload.ts
@@ -1,0 +1,16 @@
+import {ReflectTemplatePromptUpdateDescriptionPayloadResolvers} from '../resolverTypes'
+
+export type ReflectTemplatePromptUpdateDescriptionPayloadSource =
+  | {
+      promptId: string
+    }
+  | {error: {message: string}}
+
+const ReflectTemplatePromptUpdateDescriptionPayload: ReflectTemplatePromptUpdateDescriptionPayloadResolvers =
+  {
+    prompt: (source, _args, {dataLoader}) => {
+      return 'promptId' in source ? dataLoader.get('reflectPrompts').load(source.promptId) : null
+    }
+  }
+
+export default ReflectTemplatePromptUpdateDescriptionPayload
