+++ packages/server/graphql/public/types/ReflectTemplatePromptUpdateGroupColorPayload.ts
@@ -1,0 +1,16 @@
+import {ReflectTemplatePromptUpdateGroupColorPayloadResolvers} from '../resolverTypes'
+
+export type ReflectTemplatePromptUpdateGroupColorPayloadSource =
+  | {
+      promptId: string
+    }
+  | {error: {message: string}}
+
+const ReflectTemplatePromptUpdateGroupColorPayload: ReflectTemplatePromptUpdateGroupColorPayloadResolvers =
+  {
+    prompt: (source, _args, {dataLoader}) => {
+      return 'promptId' in source ? dataLoader.get('reflectPrompts').load(source.promptId) : null
+    }
+  }
+
+export default ReflectTemplatePromptUpdateGroupColorPayload
