+++ packages/server/graphql/public/types/RemoveReflectTemplatePromptPayload.ts
@@ -1,0 +1,20 @@
+import {RemoveReflectTemplatePromptPayloadResolvers} from '../resolverTypes'
+
+export type RemoveReflectTemplatePromptPayloadSource =
+  | {
+      promptId: string
+      templateId: string
+    }
+  | {error: {message: string}}
+
+const RemoveReflectTemplatePromptPayload: RemoveReflectTemplatePromptPayloadResolvers = {
+  reflectTemplate: (source, _args, {dataLoader}) => {
+    return 'templateId' in source ? dataLoader.get('meetingTemplates').load(templateId) : null
+  },
+
+  prompt: (source, _args, {dataLoader}) => {
+    return 'promptId' in source ? dataLoader.get('reflectPrompts').load(source.promptId) : null
+  }
+}
+
+export default RemoveReflectTemplatePromptPayload
